added overflow guard to softmax and moved to activations.py
